+++ src/features/residents/index.ts
@@ -9,3 +9,5 @@ export { default as ResidentsPage } from './pages/ResidentsPage';
 export { default as ResidentProfilePage } from './pages/ResidentProfilePage';
 export { default as ResidentOnboardingPage } from './pages/ResidentOnboardingPage';
 export { default as ResidentOnboardingWizard } from './components/ResidentOnboardingWizard';
+export { ResidentFinancialProfile } from '../finance/components/ResidentFinancialProfile';
+